third_party/blink/web_tests/external/wpt/fetch/private-network-access/shared-worker-fetch.tentative.window-expected.txt

This is a testharness.js-based test.
[FAIL] private to local: failure.
  assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
[FAIL] public to local: failure.
  assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
[FAIL] public to private: failure.
  assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
[FAIL] treat-as-public to local: failure.
  assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
[FAIL] treat-as-public to private: failure.
  assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
[FAIL] private https to local: failure.
  assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
[FAIL] public https to local: failure.
  assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
[FAIL] public https to private: failure.
  assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
Harness: the test ran to completion.
